move the handler to yargs-adapter to make it a complete command-module
